+++ src/AngleSharp.Js.Tests/DomTests.cs
@@ -201,6 +201,20 @@ namespace AngleSharp.Js.Tests
         }
 
         [Test]
+        public async Task QuerySelectorAllResultCanUseNodeListForEach()
+        {
+            var result = await "(function () { var nodes = document.querySelectorAll('script'); var names = []; nodes.forEach(function (node, index, collection) { names.push(index + ':' + node.nodeName + ':' + (collection === nodes)); }); return typeof nodes.forEach + ',' + names.join(); })()".EvalScriptAsync();
+            Assert.AreEqual("function,0:SCRIPT:true", result);
+        }
+
+        [Test]
+        public async Task QuerySelectorAllResultCanUseNodeListIterators()
+        {
+            var result = await "(function () { var nodes = document.querySelectorAll('script'); var entry = nodes.entries().next().value; return typeof nodes.entries + ',' + typeof nodes.keys + ',' + typeof nodes.values + ',' + nodes.keys().next().value + ',' + nodes.values().next().value.nodeName + ',' + entry[0] + ':' + entry[1].nodeName; })()".EvalScriptAsync();
+            Assert.AreEqual("function,function,function,0,SCRIPT,0:SCRIPT", result);
+        }
+
+        [Test]
         public async Task CollectionCanBeIterated()
         {
             var result = await "(function () { var n = 0; for (var s of document.getElementsByTagName('script')) { n += s.nodeName.length; } return n; })()".EvalScriptAsync();